(1) EXPERIMENTAL: Complete redefinition of frames() by duplicating much of the behaviour of delay() and ensuring proper animation based on the delay instead. The animation is now scheduled via a separate loop at the end of scheduleTick.
Seems to work much better overall, though not entirely
happy with tab switching behaviour. At least, the pattern
that is being played never gets disrupted, for example, by
one track getting offset relative to another, for instance.
.... in at least the many instances I tested it.

(2) Put in a change to steller.play whereby it will wait for the audio
context's "currentTime" to become non-zero before playing the given
model. This is necessary (at the moment) for the web audio api
to work on iOS6.