Tweaks to the scheduler main loop , delay and frames to solve the following problem -
Delay and frames were purely relying on the rate integrated
time to decide when to delay the next computation. One more
criterion is when the delay calculations can suspend for another
cycle because we're already computing ahead of time. So this commit
inserts an extra poll cycle now and then to account for the
fact that the tracks may be computing ahead to different extents.
Also, in such cases, we need to ensure that queue processing
happens at least once every requestAnimationFrame in order to
touch those suspended delays.