Behaviour change: Params mapped to sliders should use "input" event instead of "change".
Behaviour change: Params mapped to sliders should use "input" event instead of "change".

The previous code was using "change" events erroneously when instant response
to slider value changes is required to be reflected in the parameter. Now
these use the "input" event instead which is the one that is expected to give
the instant response.

The problem was that Chrome was formerly wrong about how it implemented the
"change" event, which used to fire continuously during slider value change.
This has been "fixed", which meant that steller Param sliders wouldnt change
the param's values until the mouse was released.